requirements.txt
django==3.2
djangorestframework
ipython
pdbpp
pytest
pytest-cov
pytest-django